+++ sdk/identity/identity/src/msal/utils.ts
@@ -1,17 +1,17 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import type { AuthenticationRecord, MsalAccountInfo, MsalToken, ValidMsalToken } from "./types";
-import { AuthenticationRequiredError, CredentialUnavailableError } from "../errors";
-import type { CredentialLogger } from "../util/logging";
-import { credentialLogger, formatError } from "../util/logging";
-import { DefaultAuthorityHost, DefaultTenantId } from "../constants";
+import type { AuthenticationRecord, MsalAccountInfo, MsalToken, ValidMsalToken } from "./types.js";
+import { AuthenticationRequiredError, CredentialUnavailableError } from "../errors.js";
+import type { CredentialLogger } from "../util/logging.js";
+import { credentialLogger, formatError } from "../util/logging.js";
+import { DefaultAuthorityHost, DefaultTenantId } from "../constants.js";
 import { randomUUID as coreRandomUUID, isNode, isNodeLike } from "@azure/core-util";
 
 import { AbortError } from "@azure/abort-controller";
 import type { AzureLogLevel } from "@azure/logger";
 import type { GetTokenOptions } from "@azure/core-auth";
-import { msalCommon } from "./msal";
+import { msalCommon } from "./msal.js";
 
 export interface ILoggerCallback {
   (level: msalCommon.LogLevel, message: string, containsPii: boolean): void;