+++ sdk/identity/identity/test/public/node/caeARM.spec.ts
@@ -3,10 +3,10 @@
 
 /* eslint-disable @typescript-eslint/no-non-null-asserted-optional-chain */
 
-import type { AccessToken, TokenCredential } from "../../../src";
-import { DeviceCodeCredential, UsernamePasswordCredential } from "../../../src";
-import type { MsalTestCleanup } from "../../node/msalNodeTestSetup";
-import { msalNodeTestSetup } from "../../node/msalNodeTestSetup";
+import type { AccessToken, TokenCredential } from "../../../src/index.js";
+import { DeviceCodeCredential, UsernamePasswordCredential } from "../../../src/index.js";
+import type { MsalTestCleanup } from "../../node/msalNodeTestSetup.js";
+import { msalNodeTestSetup } from "../../node/msalNodeTestSetup.js";
 import type { Recorder } from "@azure-tools/test-recorder";
 import { delay, env } from "@azure-tools/test-recorder";
 import {
@@ -15,11 +15,10 @@ import {
   createEmptyPipeline,
   createPipelineRequest,
 } from "@azure/core-rest-pipeline";
-import type { Context } from "mocha";
-import { DeveloperSignOnClientId } from "../../../src/constants";
-import { IdentityClient } from "../../../src/client/identityClient";
-import { assert } from "chai";
+import { DeveloperSignOnClientId } from "../../../src/constants.js";
+import { IdentityClient } from "../../../src/client/identityClient.js";
 import { authorizeRequestOnClaimChallenge } from "@azure/core-client";
+import { describe, it, assert, expect, vi, beforeEach, afterEach } from "vitest";
 
 /**
  * Sequence of events needed to test the CAE challenges on the Graph endpoint.
@@ -137,8 +136,8 @@ describe.skip("CAE", function () {
   let cleanup: MsalTestCleanup;
   let recorder: Recorder;
 
-  beforeEach(async function (this: Context) {
-    const setup = await msalNodeTestSetup(this.currentTest);
+  beforeEach(async function (ctx) {
+    const setup = await msalNodeTestSetup(ctx);
     cleanup = setup.cleanup;
     recorder = setup.recorder;
   });
@@ -146,7 +145,7 @@ describe.skip("CAE", function () {
     await cleanup();
   });
 
-  it("DeviceCodeCredential", async function (this: Context) {
+  it("DeviceCodeCredential", async function (ctx) {
     const [firstAccessToken, finalAccessToken] = await challengeFlow(
       new DeviceCodeCredential(recorder.configureClientOptions({ tenantId: env.AZURE_TENANT_ID })),
       recorder,
@@ -155,7 +154,7 @@ describe.skip("CAE", function () {
     assert.notDeepEqual(firstAccessToken, finalAccessToken);
   });
 
-  it("UsernamePasswordCredential", async function (this: Context) {
+  it("UsernamePasswordCredential", async function (ctx) {
     // Important: Recording this test may only work in certain tenants.
 
     const [firstAccessToken, finalAccessToken] = await challengeFlow(