+++ sdk/loadtesting/load-testing-rest/src/getTestRunCompletionPoller.ts
@@ -4,15 +4,15 @@
 import type { AbortSignalLike } from "@azure/abort-controller";
 import { AbortError } from "@azure/abort-controller";
 import type { CancelOnProgress, OperationState, SimplePollerLike } from "@azure/core-lro";
-import type { TestRunCompletionPoller, PolledOperationOptions } from "./models";
-import type { AzureLoadTestingClient } from "../generated/clientDefinitions";
+import type { TestRunCompletionPoller, PolledOperationOptions } from "./models.js";
+import type { AzureLoadTestingClient } from "../generated/clientDefinitions.js";
 import type {
   LoadTestRunCreateOrUpdateTestRun200Response,
   LoadTestRunCreateOrUpdateTestRun201Response,
   LoadTestRunGetTestRun200Response,
-} from "../generated/responses";
-import { isUnexpected } from "../generated/isUnexpected";
-import { sleep, isTestRunInProgress } from "./util/LROUtil";
+} from "../generated/responses.js";
+import { isUnexpected } from "../generated/isUnexpected.js";
+import { sleep, isTestRunInProgress } from "./util/LROUtil.js";
 
 /**
  * Creates a poller to poll for test run status.