+++ sdk/loadtesting/load-testing-rest/src/index.ts
@@ -1,16 +1,16 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import AzureLoadTesting from "../generated/azureLoadTesting";
+import AzureLoadTesting from "../generated/azureLoadTesting.js";
 
-export * from "../generated/azureLoadTesting";
-export * from "../generated/parameters";
-export * from "../generated/responses";
-export * from "../generated/clientDefinitions";
-export * from "../generated/isUnexpected";
-export * from "../generated/models";
-export * from "./models";
-export * from "../generated/outputModels";
-export * from "../generated/paginateHelper";
-export { getLongRunningPoller } from "./pollingHelper";
+export * from "../generated/azureLoadTesting.js";
+export * from "../generated/parameters.js";
+export * from "../generated/responses.js";
+export * from "../generated/clientDefinitions.js";
+export * from "../generated/isUnexpected.js";
+export * from "../generated/models.js";
+export * from "./models.js";
+export * from "../generated/outputModels.js";
+export * from "../generated/paginateHelper.js";
+export { getLongRunningPoller } from "./pollingHelper.js";
 export default AzureLoadTesting;