+++ sdk/loadtesting/load-testing-rest/test/public/testRun.spec.ts
@@ -3,30 +3,29 @@
 
 import type { Recorder } from "@azure-tools/test-recorder";
 import { isPlaybackMode } from "@azure-tools/test-recorder";
-import { assert } from "chai";
-import { createRecorder, createClient } from "./utils/recordedClient";
-import type { Context } from "mocha";
-import * as fs from "fs";
-import type { AzureLoadTestingClient } from "../../generated";
-import { isUnexpected } from "../../generated";
+import { createRecorder, createClient } from "./utils/recordedClient.js";
+import * as fs from "node:fs";
+import type { AzureLoadTestingClient } from "../../src/index.js";
+import { isUnexpected } from "../../src/index.js";
 import { isNodeLike } from "@azure/core-util";
-import { getLongRunningPoller } from "../../src/pollingHelper";
+import { getLongRunningPoller } from "../../src/pollingHelper.js";
+import { describe, it, assert, beforeEach, afterEach } from "vitest";
 
 describe("Test Run Creation", () => {
   let recorder: Recorder;
   let client: AzureLoadTestingClient;
   let readStreamTestFile: fs.ReadStream;
 
-  beforeEach(async function (this: Context) {
-    recorder = await createRecorder(this);
+  beforeEach(async (ctx) => {
+    recorder = await createRecorder(ctx);
     if (!isNodeLike || isPlaybackMode()) {
-      this.skip();
+      ctx.skip();
     }
     client = createClient(recorder);
     readStreamTestFile = fs.createReadStream("./test/public/sample.jmx");
   });
 
-  afterEach(async function () {
+  afterEach(async () => {
     await recorder.stop();
   });
 
@@ -147,10 +146,10 @@ describe("Test Run Creation", () => {
       body: {
         components: {
           "/subscriptions/{SUBSCRIPTION_ID}/resourceGroups/App-Service-Sample-Demo-rg/providers/Microsoft.Web/sites/App-Service-Sample-Demo":
-            {
-              resourceName: "App-Service-Sample-Demo",
-              resourceType: "Microsoft.Web/sites",
-            },
+          {
+            resourceName: "App-Service-Sample-Demo",
+            resourceType: "Microsoft.Web/sites",
+          },
         },
       },
     });