+++ sdk/synapse/synapse-artifacts/src/operationsInterfaces/dataFlowOperations.ts
@@ -6,9 +6,9 @@
  * Changes may cause incorrect behavior and will be lost if the code is regenerated.
  */
 
-import { PagedAsyncIterableIterator } from "@azure/core-paging";
-import { SimplePollerLike, OperationState } from "@azure/core-lro";
-import {
+import type { PagedAsyncIterableIterator } from "@azure/core-paging";
+import type { SimplePollerLike, OperationState } from "@azure/core-lro";
+import type {
   DataFlowResource,
   DataFlowGetDataFlowsByWorkspaceOptionalParams,
   DataFlowCreateOrUpdateDataFlowOptionalParams,
@@ -18,23 +18,22 @@ import {
   DataFlowDeleteDataFlowOptionalParams,
   ArtifactRenameRequest,
   DataFlowRenameDataFlowOptionalParams,
-} from "../models";
+} from "../models/index.js";
 
-/// <reference lib="esnext.asynciterable" />
 /** Interface representing a DataFlowOperations. */
 export interface DataFlowOperations {
   /**
    * Lists data flows.
-   * @param options The options parameters.
+   * @param options - The options parameters.
    */
   listDataFlowsByWorkspace(
     options?: DataFlowGetDataFlowsByWorkspaceOptionalParams,
   ): PagedAsyncIterableIterator<DataFlowResource>;
   /**
    * Creates or updates a data flow.
-   * @param dataFlowName The data flow name.
-   * @param dataFlow Data flow resource definition.
-   * @param options The options parameters.
+   * @param dataFlowName - The data flow name.
+   * @param dataFlow - Data flow resource definition.
+   * @param options - The options parameters.
    */
   beginCreateOrUpdateDataFlow(
     dataFlowName: string,
@@ -48,9 +47,9 @@ export interface DataFlowOperations {
   >;
   /**
    * Creates or updates a data flow.
-   * @param dataFlowName The data flow name.
-   * @param dataFlow Data flow resource definition.
-   * @param options The options parameters.
+   * @param dataFlowName - The data flow name.
+   * @param dataFlow - Data flow resource definition.
+   * @param options - The options parameters.
    */
   beginCreateOrUpdateDataFlowAndWait(
     dataFlowName: string,
@@ -59,8 +58,8 @@ export interface DataFlowOperations {
   ): Promise<DataFlowCreateOrUpdateDataFlowResponse>;
   /**
    * Gets a data flow.
-   * @param dataFlowName The data flow name.
-   * @param options The options parameters.
+   * @param dataFlowName - The data flow name.
+   * @param options - The options parameters.
    */
   getDataFlow(
     dataFlowName: string,
@@ -68,8 +67,8 @@ export interface DataFlowOperations {
   ): Promise<DataFlowGetDataFlowResponse>;
   /**
    * Deletes a data flow.
-   * @param dataFlowName The data flow name.
-   * @param options The options parameters.
+   * @param dataFlowName - The data flow name.
+   * @param options - The options parameters.
    */
   beginDeleteDataFlow(
     dataFlowName: string,
@@ -77,8 +76,8 @@ export interface DataFlowOperations {
   ): Promise<SimplePollerLike<OperationState<void>, void>>;
   /**
    * Deletes a data flow.
-   * @param dataFlowName The data flow name.
-   * @param options The options parameters.
+   * @param dataFlowName - The data flow name.
+   * @param options - The options parameters.
    */
   beginDeleteDataFlowAndWait(
     dataFlowName: string,
@@ -86,9 +85,9 @@ export interface DataFlowOperations {
   ): Promise<void>;
   /**
    * Renames a dataflow.
-   * @param dataFlowName The data flow name.
-   * @param request proposed new name.
-   * @param options The options parameters.
+   * @param dataFlowName - The data flow name.
+   * @param request - proposed new name.
+   * @param options - The options parameters.
    */
   beginRenameDataFlow(
     dataFlowName: string,
@@ -97,9 +96,9 @@ export interface DataFlowOperations {
   ): Promise<SimplePollerLike<OperationState<void>, void>>;
   /**
    * Renames a dataflow.
-   * @param dataFlowName The data flow name.
-   * @param request proposed new name.
-   * @param options The options parameters.
+   * @param dataFlowName - The data flow name.
+   * @param request - proposed new name.
+   * @param options - The options parameters.
    */
   beginRenameDataFlowAndWait(
     dataFlowName: string,